CSP2: Fork a draft of CSP2.
We're not done with CSP2 yet, and developing it on a branch makes it
difficult to publish a visible editor's draft. Forking it into a new
directory will be annoying, as I'll need to make sure to port changes
to the new CSP3 draft, but I'm hoping we won't end up with too many
changes at this point.