CSP2/CSP3: Rewrite path parsing.
Based on Brian and Anne's feedback, this seems more correct: we now
treat the URL's path as a list of segments, and segment the source
expression's path in the same way. We percent-decode both the URL and
the source expression as we do the comparison, which should take care of
the examples Brian has provided thus far.