Change `scheme-source` grammar
There is a small inaccuracy in source expression matching algorithm:
In section [6.1.11.4](https://api.csswg.org/bikeshed/#match-url-to-source-expression)
we hit steps 2.1.x if the expression matches the `scheme-source` or `host-source`
grammar. However, steps 2.1.x will hit only `host-source` checks, as `scheme-part` is
not defined for `scheme-source`.

This patch changes the definition of `scheme-src` to contain `scheme-part`, so the
logic in the source expression matching algorithm aligns with the intended logic.