'SecurityPolicyViolation' targeting.
If we shift the DOM event to one which bubbles, we can change their
targets to the responsible element when one is available (which is
generally the case for inline violations, and generally not the
case for Fetch-based violations).

h/t @foolip for the good idea. He also suggested that we need to think
harder about the implications for shadow DOM and fullscreen. Putting
this strawman together so we have something concrete to talk about.