Make algo 6.6.1.1 accumulate violated directives
Fix #195

This is a little clunky - there are multiple ways to express this algorithm. And it returns a comma-separated set of directives, as a random choice.

I think I've got all the spots where the effect carries through to other algorithms, but it's late, so checking twice would be smart.