Match hash-algorithm parts case-insensitively (as CSP2) (#464)
At https://w3.org/TR/CSP2/#source-list-valid-hashes the CSP2 spec says a
case-insensitive match must be performed for checking hash-algorithm
parts, and that’s how implementations actually do it.

So this change makes the current CSP spec also say that’s how it must be done.

Otherwise, without this change, the CSP spec says that the match must be
done case-sensitively. So conforming to that spec language as written
would require implementations to change their behavior in a
backward-incompatible way.