Remove `navigate-to`. (#564)
Though there's an implementation of this directive behind a flag in
Chromium, it's not something that any vendor has shipped, and there
are real concerns about information leaks that it enables. This
patch removes it from the spec to avoid confusion while we determine
what, if anything, we want to do in this space.

Partially addresses #563.